third_party/blink/web_tests/external/wpt/xhr/json.any.worker-expected.txt

This is a testharness.js-based test.
PASS Ensure the correct JSON parser is used
FAIL Ensure UTF-16 results in an error assert_equals: expected null but got object "[object Object]"
Harness: the test ran to completion.
